+++ common/src/main/java/cn/zorcc/common/network/api/Protocol.java
@@ -13,20 +13,20 @@ public interface Protocol {
      *   the parameter len will always be the exact length of reserved segment
      *   return a positive number to indicate actual bytes read, or a flag to indicate a state change
      */
-    int onReadableEvent(MemorySegment reserved, int len);
+    long onReadableEvent(MemorySegment reserved, long len);
 
     /**
      *   Indicates that protocol could write to the socket now, this function will always be invoked on poller thread
      *   return a flag to indicate a state change
      */
-    int onWritableEvent();
+    long onWritableEvent();
 
     /**
      *   Perform the actual write operation, this function will always be invoked on writer thread
      *   the parameter len will always be the exact length of data segment
      *   return a positive number to indicate actual bytes written, or a flag to indicate a state change
      */
-    int doWrite(MemorySegment data, int len);
+    long doWrite(MemorySegment data, long len);
 
     /**
      *   Perform the actual shutdown operation, this function will always be invoked on writer thread
